Add IntoIterator-enabled enumerate() and rev()
For me, the transition from `for elt in &v` to
`for (i, elt) in v.iter().enumerate()` is too jarring.

The `IntoIterator` trait transfers power back to free functions in
favour of method chaining, to some extent. With an
`IntoIterator`-enabled free function, the change is that you go from
`for elt in &v` to `for (i, elt) in enumerate(&v)`.

An alternative route could of course be to add a restricted number of
iterator adaptors directly as an extension on all `IntoIterator` types.