Replace kmerge binary heap with std library binary heap
Use PeekMut to allow for updating the tip, then doing
a sort_down when the reference is dropped, similar to how it
works now in the locally code binary heap.

Use a new 'nightly' feature PeekMut::pop() to be able to
pop the binary heap instead of sort_down.
Build with 'cargo build --features nightly',

The non-nightly form has 2 extra compares on the case where
the top entry will be popped.  The compares won't cause a move,
because the 'head' entry hasn't changed, and will still be top.

Perf issue: the kmerge_default bench test takes about 3x the time.
Copying the implementation of sort_down into the Rust std lib
got most of that perf back in a local test.