`TupleCombinations::size_hint`
`TupleCombinations` and `Tuple1Combination` merely wrap an iterator so I just call its `size_hint` method.
`Tuple2Combination` chains a wrapped `Tuple1Combination` with a smaller `Tuple2Combination` but this one is not initialized yet so it is full and therefore has `binomial(n, k)` elements.
Similar for `3..=12`, it's just recursive.