`Tuple*Combinations::fold` (macro)
`I::Item` and `A` are the same so the "where condition" is not really changed, but it is apparently needed here.
We fold `c` then for each item of `iter`, we fold the cloned `iter`.
The `while let` loop should probably be converted to `fold` somehow later (if possible) but the core logic is done and it is already really faster.