`MultiProduct`: add `MultiProductInner`
By wrapping "inner" in an option, I'll be able to fuse the iterator.

Keep the current value of each iterator in `cur`: none at the beginning, some later.
Previously, each `MultiProductIter` remembered its own element. Now, we have a vector of them. That way, we can update `cur` in-place and clone it to generate the item, I think that's simpler (less unwrap) and maybe more efficient. But we have two different vectors instead of a single bigger vector.