feat: make izip! temporary friendly
- some izip! calls with temporary values would not compile with the current
izip! implementation due to the use of a block to create the temporary iter
chain, if a user called izip! with an iterator from a temporary value the
compilation would fail
- this change adapts the macro to stop making use of the block and calls
$crate::__std_iter::Iterator::zip directly to avoid the need for temporary
iter bindings, this means the recursion produces a tuple of the form
(a, (b, (c, d))) instead of (((a, b), c), d), needing a slight adaptation
of the closure arm to invert the way arguments are unpacked, it also adds
two arms to recursively build the zipped iterators without the final map
adapter as we don't process all iterators at once as before