Implement `Debug` for remaining public types
Implement `Debug` for the `ChunkBy`-related types.  Also derive it for
two `Coalesce`-related types; I'm not sure they are actually exposed but
the compiler thinks so, and it's easy to do.

There are no more public types without a `Debug` impl (at least a
conditional one) so enable warning for missing `Debug` implementations.